requirements.txt
PyQt5>=5.15.11
pyvista>=0.46.4
pyvistaqt>=0.11.3
numpy>=1.24.0
reportlab>=4.0.0
pyinstaller>=6.0.0
requests>=2.31.0
meshio>=5.3.5
cadquery>=2.3.0
rhino3dm>=8.0.0
vtk>=9.0.0,<9.6.0
# Pygfx viewer (WebGPU, avoids Windows black screen)
pygfx
wgpu
trimesh
rtree  # Required for trimesh ray casting (annotation mode picking)
rendercanvas
scipy>=1.10.0
Pillow>=10.0.0
pillow-heif>=0.13.0  # HEIC (iPhone) to JPEG conversion in annotation mode
# aspose-3d>=24.0.0  # Optional: PDF 3D export. Not needed for STL loading. Uncomment if required.
